Refaktorerer til en felles gyldig søknad og endrer der det er behov. Slipper x antall fulle json body
